Add files via upload
The motivation for this project, focused on Homonyms, arises from the persistent challenges encountered in natural language processing (NLP), particularly with respect to nuanced homonym interpretation in English. Homonyms, words with shared spellings or pronunciations but distinct meanings, introduce ambiguity in language processing tasks. This complexity is compounded by the intricacies of the English language, posing obstacles to accurate homonym disambiguation.